feat(events): stream canceled/deleted (many) events (#1272)

+++ v2/apiserver/internal/core/kubernetes/substrate.go
@@ -643,40 +643,45 @@ func (s *substrate) DeleteWorkerAndJobs(
 		},
 	).AsSelector().String()
 
-	// Delete all pods related to this Event
-	if err := s.kubeClient.CoreV1().Pods(
-		project.Kubernetes.Namespace,
-	).DeleteCollection(
-		ctx,
-		metav1.DeleteOptions{},
-		metav1.ListOptions{
-			LabelSelector: labelSelector,
-		},
-	); err != nil {
-		return errors.Wrapf(
-			err,
-			"error deleting event %q pods in namespace %q",
-			event.ID,
+	// If a worker's phase is CANCELED or PENDING, no pods would have been
+	// created. Therefore, we just skip to cleaning up the event secret(s) below.
+	if event.Worker.Status.Phase != core.WorkerPhaseCanceled &&
+		event.Worker.Status.Phase != core.WorkerPhasePending {
+		// Delete all pods related to this Event
+		if err := s.kubeClient.CoreV1().Pods(
 			project.Kubernetes.Namespace,
-		)
-	}
+		).DeleteCollection(
+			ctx,
+			metav1.DeleteOptions{},
+			metav1.ListOptions{
+				LabelSelector: labelSelector,
+			},
+		); err != nil {
+			return errors.Wrapf(
+				err,
+				"error deleting event %q pods in namespace %q",
+				event.ID,
+				project.Kubernetes.Namespace,
+			)
+		}
 
-	// Delete all persistent volume claims related to this Event
-	if err := s.kubeClient.CoreV1().PersistentVolumeClaims(
-		project.Kubernetes.Namespace,
-	).DeleteCollection(
-		ctx,
-		metav1.DeleteOptions{},
-		metav1.ListOptions{
-			LabelSelector: labelSelector,
-		},
-	); err != nil {
-		return errors.Wrapf(
-			err,
-			"error deleting event %q persistent volume claims in namespace %q",
-			event.ID,
+		// Delete all persistent volume claims related to this Event
+		if err := s.kubeClient.CoreV1().PersistentVolumeClaims(
 			project.Kubernetes.Namespace,
-		)
+		).DeleteCollection(
+			ctx,
+			metav1.DeleteOptions{},
+			metav1.ListOptions{
+				LabelSelector: labelSelector,
+			},
+		); err != nil {
+			return errors.Wrapf(
+				err,
+				"error deleting event %q persistent volume claims in namespace %q",
+				event.ID,
+				project.Kubernetes.Namespace,
+			)
+		}
 	}
 
 	// Delete all secrets related to this Event
